+++ src/main/java/org/cryptomator/common/mount/MountModule.java
@@ -2,9 +2,12 @@ package org.cryptomator.common.mount;
 
 import dagger.Module;
 import dagger.Provides;
+import org.cryptomator.common.ObservableUtil;
+import org.cryptomator.common.settings.Settings;
 import org.cryptomator.integrations.mount.MountService;
 
 import javax.inject.Singleton;
+import javafx.beans.value.ObservableValue;
 import java.util.List;
 
 @Module
@@ -14,6 +17,16 @@ public class MountModule {
 	@Singleton
 	static List<MountService> provideSupportedMountServices() {
 		return MountService.get().toList();
+	}
+
+	@Provides
+	@Singleton
+	static ObservableValue<MountService> provideDefaultMountService(List<MountService> mountProviders, Settings settings) {
+		var fallbackProvider = mountProviders.stream().findFirst().orElse(null);
+
+		return ObservableUtil.mapWithDefault(settings.mountService, //
+				serviceName -> mountProviders.stream().filter(s -> s.getClass().getName().equals(serviceName)).findFirst().orElse(fallbackProvider), //
+				fallbackProvider);
 	}
 
 }
